move 'windows bin linking shim should work' to its own file
